requirements.txt
# packages in environment at /home/ejnik/anaconda3/envs/adist:
#
# Name                    Version                   Build  Channel
apbs                      1.5                  h14c3975_3    schrodinger
asn1crypto                0.24.0                   py36_3    conda-forge
bzip2                     1.0.6                h470a237_2    conda-forge
ca-certificates           2018.8.24            ha4d7672_0    conda-forge
certifi                   2018.8.24                py36_1    conda-forge
cffi                      1.11.5           py36h5e8e0c9_1    conda-forge
chardet                   3.0.4                    py36_3    conda-forge
cryptography              2.3.1            py36hdffb7b8_0    conda-forge
cryptography-vectors      2.3.1                    py36_0    conda-forge
cycler                    0.10.0                     py_1    conda-forge
dbus                      1.13.0               h3a4f0e9_0    conda-forge
expat                     2.2.5                hfc679d8_2    conda-forge
fontconfig                2.13.0               hd36ec8e_5    conda-forge
freemol                   1.158                    py36_1    schrodinger
freetype                  2.8.1                hfa320df_1    conda-forge
gettext                   0.19.8.1                      0    conda-forge
glew                      2.0.0                         0    schrodinger
glib                      2.55.0               h464dc38_2    conda-forge
gst-plugins-base          1.12.5               hde13a9d_0    conda-forge
gstreamer                 1.12.5               h61a6719_0    conda-forge
h5py                      2.8.0            py36hb794570_1    conda-forge
hdf5                      1.10.2               hc401514_1    conda-forge
icu                       58.2                 hfc679d8_0    conda-forge
idna                      2.7                      py36_2    conda-forge
jpeg                      9c                   h470a237_1    conda-forge
kiwisolver                1.0.1            py36h2d50403_2    conda-forge
libffi                    3.2.1                hfc679d8_4    conda-forge
libgcc-ng                 7.2.0                hdf63c60_3    conda-forge
libgfortran               3.0.0                         1    conda-forge
libgfortran-ng            7.2.0                hdf63c60_3    conda-forge
libiconv                  1.15                 h470a237_3    conda-forge
libopenblas               0.2.20               h9ac9557_7  
libpng                    1.6.34               ha92aebf_2    conda-forge
libstdcxx-ng              7.2.0                hdf63c60_3    conda-forge
libtiff                   4.0.9                he6b73bb_1    conda-forge
libuuid                   2.32.1               h470a237_0    conda-forge
libxcb                    1.13                 h470a237_2    conda-forge
libxml2                   2.9.8                h422b904_3    conda-forge
matplotlib                2.2.2                    py36_1    conda-forge
mengine                   1                    h14c3975_1    schrodinger
mpeg_encode               1                    h14c3975_1    schrodinger
msgpack-python            0.5.6            py36h2d50403_2    conda-forge
mtz2ccp4_px               1.0                  h9ac9557_2    schrodinger
ncurses                   6.1                  hfc679d8_1    conda-forge
numpy                     1.14.3           py36h28100ab_1  
numpy-base                1.14.3           py36h0ea5e3f_1  
olefile                   0.45.1                     py_1    conda-forge
openssl                   1.0.2p               h470a237_0    conda-forge
pcre                      8.41                 h470a237_2    conda-forge
pdb2pqr                   2.1.1                    py36_1    schrodinger
pillow                    5.2.0            py36h2dc6135_1    conda-forge
pip                       18.0                     py36_1    conda-forge
pmw                       2.0.1                    py36_2    schrodinger
pthread-stubs             0.4                  h470a237_1    conda-forge
pycparser                 2.18                       py_1    conda-forge
pymol                     2.2.0            py36hf484d3e_0    schrodinger
pyopenssl                 18.0.0                   py36_0    conda-forge
pyparsing                 2.2.0                      py_1    conda-forge
pyqt                      5.6.0            py36h8210e8a_7    conda-forge
pysocks                   1.6.8                    py36_2    conda-forge
python                    3.6.6                h5001a0f_0    conda-forge
python-dateutil           2.7.3                      py_0    conda-forge
pytz                      2018.5                     py_0    conda-forge
qt                        5.6.2                h50c60fd_8    conda-forge
readline                  7.0                  haf1bffa_1    conda-forge
requests                  2.19.1                   py36_1    conda-forge
rigimol                   1.3                           2    schrodinger
setuptools                40.2.0                   py36_0    conda-forge
simplemmtf-python         0.1                      py36_1    schrodinger
sip                       4.18.1           py36hfc679d8_0    conda-forge
six                       1.11.0                   py36_1    conda-forge
sqlite                    3.24.0               h2f33b56_1    conda-forge
tk                        8.6.8                         0    conda-forge
tornado                   5.1              py36h470a237_1    conda-forge
urllib3                   1.23                     py36_1    conda-forge
wheel                     0.31.1                   py36_1    conda-forge
xorg-libxau               1.0.8                h470a237_6    conda-forge
xorg-libxdmcp             1.1.2                h470a237_7    conda-forge
xz                        5.2.4                h470a237_1    conda-forge
zlib                      1.2.11               h470a237_3    conda-forge
